admin/requirements.txt
adrf==0.1.9
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-property==0.2.2
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.4
djangorestframework==3.15.2
frozenlist==1.5.0
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.4.2
magic-filter==1.0.12
multidict==6.1.0
pillow==11.1.0
prompt_toolkit==3.0.48
propcache==0.2.1
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pyOpenSSL==25.0.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
rsa==4.9
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
Telethon==1.38.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==14.1
yarl==1.18.3
zope.interface==7.2
